% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{deprecate}
\alias{deprecate}
\alias{mgnify_client}
\alias{mgnify_query}
\alias{mgnify_analyses_from_samples}
\alias{mgnify_analyses_from_studies}
\alias{mgnify_get_download_urls}
\alias{mgnify_download}
\alias{mgnify_get_analyses_results}
\alias{mgnify_get_analyses_phyloseq}
\alias{mgnify_get_analyses_metadata}
\alias{mgnify_retrieve_json}
\title{These functions will be deprecated. Please use other functions instead.}
\usage{
mgnify_client(
  username = NULL,
  password = NULL,
  usecache = FALSE,
  cache_dir = NULL,
  warnings = FALSE,
  use_memcache = FALSE,
  ...
)

mgnify_query(
  client,
  qtype = "samples",
  accession = NULL,
  asDataFrame = TRUE,
  maxhits = 200,
  usecache = FALSE,
  ...
)

mgnify_analyses_from_samples(client, accession, usecache = TRUE, ...)

mgnify_analyses_from_studies(client, accession, usecache = TRUE, ...)

mgnify_get_download_urls(
  client,
  accessions,
  accession_type,
  usecache = TRUE,
  ...
)

mgnify_download(
  client,
  url,
  file = NULL,
  read_func = NULL,
  usecache = TRUE,
  Debug = FALSE,
  ...
)

mgnify_get_analyses_results(
  client = NULL,
  accessions,
  retrievelist = c(),
  compact_results = TRUE,
  usecache = TRUE,
  bulk_dl = FALSE,
  ...
)

mgnify_get_analyses_phyloseq(
  client = NULL,
  accessions,
  usecache = TRUE,
  returnLists = FALSE,
  tax_SU = "SSU",
  get_tree = FALSE,
  ...
)

mgnify_get_analyses_metadata(client, accessions, usecache = TRUE, ...)

mgnify_retrieve_json(
  client,
  path = "biomes",
  complete_url = NULL,
  qopts = NULL,
  maxhits = 200,
  usecache = FALSE,
  Debug = FALSE,
  ...
)
}
\arguments{
\item{username}{-}

\item{password}{-}

\item{usecache}{-}

\item{cache_dir}{-}

\item{warnings}{-}

\item{use_memcache}{-}

\item{...}{-}

\item{client}{-}

\item{qtype}{-}

\item{accession}{-}

\item{asDataFrame}{-}

\item{maxhits}{-}

\item{accessions}{-}

\item{accession_type}{-}

\item{url}{-}

\item{file}{-}

\item{read_func}{-}

\item{Debug}{-}

\item{retrievelist}{-}

\item{compact_results}{-}

\item{bulk_dl}{-}

\item{returnLists}{-}

\item{tax_SU}{-}

\item{get_tree}{-}

\item{path}{-}

\item{complete_url}{-}

\item{qopts}{-}
}
\value{
-
}
\description{
These functions will be deprecated. Please use other functions instead.
}
