% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnalysisAccessions.R
\name{getAnalysisAccessions}
\alias{getAnalysisAccessions}
\alias{getAnalysisAccessions,MgnifyClient-method}
\title{Look up analysis accession IDs for one or more study or sample accessions}
\usage{
getAnalysisAccessions(x, type, accession, use.cache = TRUE, ...)

\S4method{getAnalysisAccessions}{MgnifyClient}(x, type, accession, use.cache = TRUE, ...)
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{type}{A single character value specifying a type of
accession IDs specified by \code{accession}. Must be "study" or "sample".}

\item{accession}{A single character value or a vector of character values
specifying study or sample accession IDs that are used to retrieve analyses
IDs.}

\item{use.cache}{A single boolean value to specify whether to re-use/store
data on disk, rather than query the server.
(By default: \code{use.cache = TRUE})}

\item{...}{Optional arguments; not currently used.}
}
\value{
vector of analysis accession IDs.
}
\description{
Look up analysis accession IDs for one or more study or sample accessions
}
\details{
Retrieve Analysis accession IDs associated with the supplied study or
sample accession.
}
\examples{
\dontrun{
# Retrieve all analysis ids from studies
# MGYS00005058, MGYS00005058 and MGYS00005058
result <- getAnalysisAccessions(myclient, "study", c("MGYS00005058"))

# Retrieve all analysis ids from samples
result <- getAnalysisAccessions(
    myclient, "sample", c("SRS4392730", "SRS4392743"))
}

}
