% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/getData.R
\name{getData}
\alias{getData}
\alias{getData,MgnifyClient-method}
\title{Versatile function to retrieve results from MGnify database}
\usage{
getData(x, ...)

\S4method{getData}{MgnifyClient}(x, type, accession.type = NULL, accession = NULL, ...)
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{...}{optional arguments:}

\item{accession}{A single character value or a vector of character values
specifying accession IDs to return results for.}
}
\description{
Versatile function to retrieve results from MGnify database
}
\examples{
# Create a client object
mg <- MgnifyClient(useCache = FALSE)

}
