% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_analyses_from_samples.R
\name{mgnify_analyses_from_samples}
\alias{mgnify_analyses_from_samples}
\title{Look up analysis accession IDs for one or more sample accessions}
\usage{
mgnify_analyses_from_samples(client, accession, usecache = T)
}
\arguments{
\item{client}{\code{mgnify_client} instance}

\item{accession}{Single sample accession id, or vector/list of accessions for which to retrieve Analyses ids}

\item{usecache}{Flag to determine whether to re-use/store data on disk, rather than query the server.}
}
\value{
vector of associated Analysis accession ids
}
\description{
\code{mgnify_analyses_from_samples} Retrieve Analysis accession IDs associated with the supplied Sample accession
}
\details{
Helper function to get all analyses associated with the given samples.
}
\examples{
#Retrieve all analysis ids from samples
result <- mgnify_analyses_from_samples(myclient, c("SRS4392730", "SRS4392743"))

}
