% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_analyses_from_studies}
\alias{mgnify_analyses_from_studies}
\title{Look up analysis accession IDs for one or more study accessions}
\usage{
mgnify_analyses_from_studies(client, accession, usecache = T)
}
\arguments{
\item{client}{\code{mgnify_client} instance}

\item{accession}{Single study accession id, or vector/list of accessions for which to retrieve Analyses ids}

\item{usecache}{Flag to determine whether to re-use/store data on disk, rather than query the server.}
}
\value{
vector of Analysis accession ids
}
\description{
\code{mgnify_analyses_from_studies} Retrieve Analysis accession IDs associated with the supplied Study accession
}
\details{
Helper function to get all analyses associated with the given studies.
}
\examples{
#Retrieve all analysis ids from studies MGYS00005058, MGYS00005058 and MGYS00005058
result <- mgnify_analyses_from_studies(myclient, c("MGYS00005058", "MGYS00005058" and "MGYS00005058"))

}
