% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_attr_list_to_df}
\alias{mgnify_attr_list_to_df}
\title{Coverting attribute lists to a single data.frame row}
\usage{
mgnify_attr_list_to_df(json, metadata_key = NULL)
}
\arguments{
\item{json}{The \emph{raw} result list}

\item{metadata_key}{Optional extra key to parse child entries and include in the output}
}
\value{
1xn data.frame
}
\description{
\code{mgnify_attr_list_to_df} extracts the \code{attribute} entry in a JSONAPI result
}
