% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_client}
\alias{mgnify_client}
\title{Instantiate the MGnifyR client object}
\usage{
mgnify_client(url = NULL, username = NULL, password = NULL,
  usecache = F, cache_dir = NULL, warnings = F, use_memcache = F)
}
\arguments{
\item{username}{optional username to authenticate.}

\item{password}{optional password for authentication.}

\item{usecache}{whether to enable on-disk caching of results during this session. In most use cases should be TRUE.}

\item{cache_dir}{specifies a folder to contain the local cache. If NULL, and usecache is TRUE, the new subdirectory \code{.MGnifyR_cache}
in the current working directory will be used. Note that cached files are persistent, so the cache directory may be reused between sessions,
taking advantage of previously downloaded results. The directory will be created if it doesn't exist already.}

\item{warnings}{debug flag to print extra output during invocation of some MGnifyR functions. Defaults to FALSE.}

\item{use_memcache}{flag to indicate whether functional results obtained when \code{bulk_dl} is \code{TRUE} in \code{mgnify_get_analyses_results} should
be stored in an in-memory cache, rather than the cached input being re-read for each accession. this is currently NOT working
properly and should therefore be set \code{FALSE} (the default). It has the potential to speed up searches considerably though, especially
for studies with a large number of samples, so will be implemented properly in the future.}
}
\description{
All functions in the MGnifyR package take a \code{mgnify_client} object as their first argument. While not essential
to querying the raw MGnify API (which is exposed as relative standard JSONAPI), the object allows the simple handling of both
user authentication and access to private data, and local on-disk caching of results.
}
\examples{
my_client <- mgnify_client(username="Webin-1122334", password="SecretPassword", usecache=T, cache_dir = "/scratch/MGnify_cache_location")
}
