% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_get_analyses_metadata}
\alias{mgnify_get_analyses_metadata}
\title{Get all Study, Sample and Analysis metadata for the supplied analyses accessions}
\usage{
mgnify_get_analyses_metadata(client, accessions, usecache = T)
}
\arguments{
\item{client}{\code{mgnify_client} instance}

\item{accessions}{Single value or list/vector of Anlysis accessions to retrieve data for}

\item{usecache}{Whether to use the disk based cache.}
}
\value{
\code{data.frame} of metadta for each analysis in the \code{accession} list.
}
\description{
\code{mgnify_get_analyses_metadata} retrieves all associated Study, Sample and Analysis metadata attributes
a list of Analyses accessions (determined from \code{mgnify_analyses_from_x})
}
\examples{

}
