% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_get_analyses_phyloseq}
\alias{mgnify_get_analyses_phyloseq}
\title{Retrieve OTU tables for all specified accessions and build a\code{phyloseq} object}
\usage{
mgnify_get_analyses_phyloseq(client = NULL, accessions, usecache = T,
  returnLists = F, tax_SU = "SSU", get_tree = FALSE)
}
\arguments{
\item{client}{\code{mgnify_client} instance}

\item{accessions}{Single value or list/vector of Analysis accessions to retrieve data for.}

\item{usecache}{Whether to use the disk based cache.}

\item{returnLists}{Flag to determine whether to merge a per-analysis \code{phyloseq} objects into one, or
return a list of single-sample objects. Since perculiarities in OTU format may lead to failed merges, returning
a list of single objects may aid debugging. In most cases though the desired behaviour is to return a single
object (returnLists = F)}

\item{tax_SU}{Which taxa subunit results should be selected? Currently, taxonomy assignments in the
MGnify pipelines rely on rRNA matches to existing databases (GreenGenes and SILVA), with later
pipelines checking both the SSU and LSU portions of the rRNA sequence. \code{tax_SU} allows the
selection of either the Small subunit (SSU) or Large subunit results in the final \code{phyloseq} object.
Older pipeline versions do not report results for both subunits,
and thus for some accessions this value will have no effect.}

\item{get_tree}{Flag to control whether to include available phylogenetic trees in the phyloseq object. At present this option is of limited use as
the trees available are specific to each accession, holding only a subset of the leaves and branches of the full canonical tree used for the pipeline. This means that
\code{phyloseq} is unable to merge the trees, and therefore fails to build a final combined object. Setting \code{returnLists} to TRUE allows the results to be returned successfully,
albeit not in a single object.}
}
\description{
\code{mgnify_get_analyses_phyloseq} retrieves all associated Study, Sample and Analysis metadata attributes,
along with all OTU tables (of a given taxonomic type), and merges them together to build a \code{phyloseq}
object with \code{otu_table}, \code{tax_table} and \code{sample_data} objects.
#'
}
\examples{


}
