% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_get_analyses_results.R
\name{mgnify_get_analyses_results}
\alias{mgnify_get_analyses_results}
\title{Get functional and/or taxonomic information for a list of accessions}
\usage{
mgnify_get_analyses_results(
  client = NULL,
  accessions,
  retrievelist = c(),
  compact_results = T,
  usecache = T,
  bulk_dl = F
)
}
\arguments{
\item{client}{a valid \code{mgnify_client} object}

\item{accessions}{list or vector of accessions to return results for}

\item{retrievelist}{list or vector of functional analysis types to retrieve, or "all" to get all available results. The current list of available
types can be found using \code{ names(MGnifyR::analyses_results_type_parsers)}. Note that not depending on the particular analysis type, pipeline
version etc., not all functional results will be available.}

\item{compact_results}{optional parameter to return a named list (one entry per element in \code{retrievelist}) of data.frames, with each data.frame
containing results for all requested accessions. If \code{FALSE}, \code{mgnify_get_analyses_results} returns a list of lists, each element consiting of
results for a single accession.}

\item{usecache}{Whether to use the MGnify local caching system to speed up searching. It is highly recommended that this be enabled (default=TRUE)}

\item{bulk_dl}{should MGnifyR attempt to speed things up by downloading relevant studies TSV results and only extracting the required columns, rather than using
the JSONAPI interface. When getting results where multiple accessions share the same study, this option may result in significantly faster processing. However, there
appear to be (quite a few) cases in the database where the TSV result columns do NOT match the expected accession names. This will hopefully be fixed in the future, but for
now \code{bulk_dl} defaults to FALSE. When it does work, it can be orders of magnitude more efficient.}

\item{analyses_results_type_parsers}{(To be described)}
}
\value{
Named list of \code{data.frames}, corresponding to the requested analysis types in \code{retrievelist}
}
\description{
Given a set of analysis accessions and collection of annotation types, \code{mgnify_get_analyses_results} queries the MGNify API
and returns the results, by default merging the results into multi-accession data.frames
}
