% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_get_download_urls}
\alias{mgnify_get_download_urls}
\title{Listing files available for download}
\usage{
mgnify_get_download_urls(client, accessions, accession_type,
  usecache = T)
}
\arguments{
\item{client}{valid MGnify client object}

\item{accessions}{list of accessions to query}

\item{accession_type}{one of \code{analysis},\code{samples},\code{studies},\code{assembly},\code{genome} or \code{run}}

\item{usecache}{whether to use the on-disk cache to speed up queries (default T)}
}
\value{
\code{data.frame} containing all discovered downloads. If multiple \code{accessions} are queried, the \code{accessions} column
may to filter the results - since rownames are not set (and wouldn;'t make sense as each query will return multiple items)
}
\description{
\code{mgnify_get_downloads} is a wrapper function allowing easy enumeration of downloads available for a given
accession (or list thereof). Returns a single data.frame containing all available downloads and associated metadata,
including the url location and description. This can then be filtered to extract the urls of interest, before actually
retrieving the files using \code{mgnify_download}
}
\examples{
#Make a client ibject
mg <- mgnify_client(cache_dir="/tmp/mgcache")
#create a vector of accession ids - these happen to be \\code{analysis} accessions
accession_vect <- c("MGYA00563876", "MGYA00563877", "MGYA00563878", "MGYA00563879", "MGYA00563880" )
downloads <- mgnify_get_downloads(mg, accession_vect, "analyses")
}
