% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_get_runs_as_phyloseq}
\alias{mgnify_get_runs_as_phyloseq}
\title{Retrieve \code{analyses} from MGnify and convert into \code{phyloseq} objects.}
\usage{
mgnify_get_runs_as_phyloseq(client = NULL, accessions = NULL,
  downloadDIR = "./tmpdownloads", use_downloads = T, pipeline_version,
  usecache = T)
}
\arguments{
\item{client}{MGnifyR client}

\item{accessions}{data.frame containing (at least) two columns: \code{accession} - MGnify accession id (sample, run, study etc), and
\code{type} - type of accession in \code{accession} column. Lists of accessions will be acceptable in the future. \code{type} defaults to \code{samples}
if type column is absent. In most cases, the \code{accession} input will be the output of a previous \code{mgnify_query} call.}

\item{downloadDIR}{Location to store retrieved \code{.biom} files. Can be resued between session to reduce load and speed up analysis.}

\item{use_downloads}{Try to use previously retrieved \code{.biom} files, or start from scratch.}

\item{pipeline_version}{Unimplemented - filtering ~should~ be done at the data.frame level, before we get to this point?}

\item{usecache}{Should JSON API queries be cached locally for performance?}
}
\value{
\code{phyloseq} object with filled \code{sample_data} and \code{otu_table} slots.
}
\description{
\code{mgnify_get_runs_as_phyloseq} takes as input a previously determined data.frame or list from \code{mgnify_query}, and downloads all associated
analysis data (principally \code{.biom} OTU table output), before merging it with corresponding \code{sample}, \code{study} and \code{run} data into a
single \code{phyloseq} object. The \code{sample_data} in the resulting phyloseq object includes all \code{attributes} metadata from corresponding \code{
samples}, \code{studies} and \code{runs} entries
}
\examples{

}
