% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgnify_get_single_analysis_treese.R
\name{mgnify_get_single_analysis_treese}
\alias{mgnify_get_single_analysis_treese}
\title{Get a single biom file and convert it to TreeSummarizedExperiment format}
\usage{
mgnify_get_single_analysis_treese(
  client = NULL,
  accession,
  usecache = T,
  downloadDIR = NULL,
  tax_SU = "SSU",
  get_tree = FALSE
)
}
\description{
Get a single biom file and convert it to TreeSummarizedExperiment format
}
