% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allAccessors.R
\name{MgnifyClient-accessors}
\alias{MgnifyClient-accessors}
\alias{url,MgnifyClient-method}
\alias{authTok,MgnifyClient-method}
\alias{cacheDir,MgnifyClient-method}
\alias{warnings,MgnifyClient-method}
\alias{useMemCache,MgnifyClient-method}
\alias{memCache,MgnifyClient-method}
\alias{clearCache,MgnifyClient-method}
\alias{url<-,MgnifyClient-method}
\alias{authTok<-,MgnifyClient-method}
\alias{cacheDir<-,MgnifyClient-method}
\alias{warnings<-,MgnifyClient-method}
\alias{useMemCache<-,MgnifyClient-method}
\alias{memCache<-,MgnifyClient-method}
\alias{clearCache<-,MgnifyClient-method}
\title{MgnifyClient accessors and mutators}
\usage{
\S4method{url}{MgnifyClient}(x)

\S4method{authTok}{MgnifyClient}(x)

\S4method{cacheDir}{MgnifyClient}(x)

\S4method{warnings}{MgnifyClient}(x)

\S4method{useMemCache}{MgnifyClient}(x)

\S4method{memCache}{MgnifyClient}(x)

\S4method{clearCache}{MgnifyClient}(x)

\S4method{url}{MgnifyClient}(x) <- value

\S4method{authTok}{MgnifyClient}(x) <- value

\S4method{cacheDir}{MgnifyClient}(x) <- value

\S4method{warnings}{MgnifyClient}(x) <- value

\S4method{useMemCache}{MgnifyClient}(x) <- value

\S4method{memCache}{MgnifyClient}(x) <- value

\S4method{clearCache}{MgnifyClient}(x) <- value
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{value}{A value to be added to a certain slot.}
}
\value{
A value of MgnifyClient object or nothing.
}
\description{
MgnifyClient accessors and mutators
}
\details{
These functions are for fetching and mutating slots of
\code{MgnifyClient} object.
}
\examples{
mg <- MgnifyClient()

url(mg)
warnings(mg) <- FALSE

}
