% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MgnifyClient.R, R/allClasses.R
\docType{class}
\name{MgnifyClient}
\alias{MgnifyClient}
\alias{MgnifyClient-class}
\title{Constructor for creating a MgnifyClient object to allow the access to
MGnify database.}
\usage{
MgnifyClient(
  username = NULL,
  password = NULL,
  useCache = FALSE,
  cacheDir = NULL,
  warnings = FALSE,
  useMemCache = FALSE,
  clearCache = FALSE,
  ...
)
}
\arguments{
\item{username}{A single character value specifying an optional username for
authentication. (By default: \code{username = NULL})}

\item{password}{A single character value specifying an optional password for
authentication. (By default: \code{password = NULL})}

\item{useCache}{A single boolean value specifying whether to enable on-disk
caching of results during this session. In most use cases should be TRUE.
(By default: \code{useCache = FALSE})}

\item{cacheDir}{A single character value specifying a folder to contain the
local cache. If NULL, and useCache is TRUE, the new subdirectory
\code{.MGnifyR_cache} in the current working directory will be used. Note
that cached files are persistent, so the cache directory may be reused
between sessions, taking advantage of previously downloaded results. The
directory will be created if it doesn't exist already.
(By default: \code{cachedir = NULL})}

\item{warnings}{A single boolean value specifying whether to print extra
output during invocation of some MGnifyR functions.
(By default: \code{warnings = FALSE})}

\item{useMemCache}{A single boolean value specifying whether to indicate
whether functional results obtained when \code{bulk_dl} is \code{TRUE}
in \code{mgnify_get_analyses_results} should be stored in an in-memory
cache, rather than the cached input being re-read for each accession. this
is currently NOT working properly and should therefore be set \code{FALSE}.
It has the potential to speed up searches considerably though, especially
for studies with a large number of samples, so will be implemented properly
in the future. (By default: \code{useMemCache = FALSE})}

\item{clearCache}{A single boolean value specifying whether to clear the
cache. (By default: \code{clearCache = FALSE})}

\item{...}{optional arguments:
\itemize{
  \item{url}{ A single character value specifying an url address of the
  the database.
  (By default: \code{url = "https://www.ebi.ac.uk/metagenomics/api/v1"})}
}}
}
\value{
A MgnifyClient object.
}
\description{
Constructor for creating a MgnifyClient object to allow the access to
MGnify database.

A MgnifyClient object
}
\details{
All functions in the MGnifyR package take a \code{MgnifyClient} object as
their first argument. While not essential to querying the raw MGnify API
(which is exposed as relative standard JSONAPI), the object allows the
simple handling of both user authentication and access to private data,
and local on-disk caching of results.

An object that are required by functions of MGnifyR package.
}
\section{Slots}{

\describe{
\item{\code{url}}{A single character value specifying an URL address of database.}

\item{\code{authTok}}{A single character value specifying authentication token.}

\item{\code{useCache}}{A single boolean value specifying whether to use cache.}

\item{\code{cacheDir}}{A single character value specifying cache directory.}

\item{\code{warnings}}{A single boolean value specifying whether to show warnings.}

\item{\code{useMemCache}}{A single boolean value specifying whether to use on-disk
memory.}

\item{\code{memCache}}{A single character value specifying on-disk memory directory.}

\item{\code{clearCache}}{A single boolean value specifying whether to clear cache.}

\item{\code{verbose}}{A single boolean value specifying whether to show messages.}
}}

\section{Constructor}{

See  \code{\link{MgnifyClient}} for constructor.
}

\section{Accessor}{

See \code{\link{MgnifyClient-accessors}} for accessor functions.
}

\examples{
my_client <- MgnifyClient(
    useCache = TRUE, cacheDir = "/scratch/MGnify_cache_location"
    )

\donttest{
# Use username and password to get access to non-public data
my_client <- MgnifyClient(
    username = "Webin-1122334", password = "SecretPassword",
    useCache = TRUE, cacheDir = "/scratch/MGnify_cache_location"
    )
}

}
