% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\docType{package}
\name{MGnifyR-package}
\alias{MGnifyR-package}
\alias{MGnifyR}
\title{\code{MGnifyR} Package.}
\description{
\code{MGnifyR} implements an interface to the EBI MGnify database.
See the vignette for a general introduction to this package.
\href{https://www.ebi.ac.uk/metagenomics/about}{About MGnify} for general MGnify
information, and
\href{https://emg-docs.readthedocs.io/en/latest/api.html}{API documentation}
for details about the JSONAPI implementation,
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
\author{
\strong{Maintainer}: Tuomas Borman \email{tuomas.v.borman@utu.fi} (\href{https://orcid.org/0000-0002-8563-8884}{ORCID})

Authors:
\itemize{
  \item Ben Allen \email{ben.allen@ncl.ac.uk}
  \item Leo Lahti \email{leo.lahti@iki.fi} (\href{https://orcid.org/0000-0001-5537-637X}{ORCID})
}

}
