% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAnalysis.R
\name{searchAnalysis}
\alias{searchAnalysis}
\alias{searchAnalysis,MgnifyClient-method}
\title{Look up analysis accession IDs for one or more study or sample accessions}
\usage{
searchAnalysis(x, type, accession, use.cache = TRUE, ...)

\S4method{searchAnalysis}{MgnifyClient}(x, type, accession, use.cache = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{type}{A single character value specifying a type of
accession IDs specified by \code{accession}. Must be "studies" or "samples".}

\item{accession}{A single character value or a vector of character values
specifying study or sample accession IDs that are used to retrieve analyses
IDs.}

\item{use.cache}{A single boolean value to specify whether to re-use/store
data on disk, rather than query the server.
(By default: \code{use.cache = TRUE})}

\item{...}{Optional arguments; not currently used.}

\item{verbose}{A single boolean value to specify whether to show
the progress bar. (By default: \code{verbose = TRUE})}
}
\value{
vector of analysis accession IDs.
}
\description{
Look up analysis accession IDs for one or more study or sample accessions
}
\details{
Retrieve analysis accession IDs associated with the supplied study or
sample accession.
}
\examples{
\dontrun{
# Retrieve all analysis ids from studies
# MGYS00005058, MGYS00005058 and MGYS00005058
result <- searchAnalysis(myclient, "study", c("MGYS00005058"))

# Retrieve all analysis ids from samples
result <- searchAnalysis(
    myclient, "sample", c("SRS4392730", "SRS4392743"))
}

}
