% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getResult.R
\name{getResult}
\alias{getResult}
\alias{getResult,MgnifyClient-method}
\title{Get functional and/or taxonomic information for a list of accessions}
\usage{
getResult(
  x,
  accession,
  get.taxa = TRUE,
  get.func = TRUE,
  output = "TreeSE",
  use.cache = TRUE,
  verbose = TRUE,
  ...
)

\S4method{getResult}{MgnifyClient}(
  x,
  accession,
  get.taxa = TRUE,
  get.func = TRUE,
  output = "TreeSE",
  use.cache = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{accession}{A single character value or a vector of character values
specifying accession IDs to return results for.}

\item{get.taxa}{A boolean value specifying whether to retrieve metagenomic
data. (By default: \code{get.taxa = TRUE})}

\item{get.func}{A boolean value or a single character value or a vector
character values specifying functional analysis types to retrieve. If
\code{get.func = TRUE}, all available functional datatypes are retrieved,
and if \code{FALSE}, functional data is not retrieved. The current list of
available types is "antismash-gene-clusters", "go-slim", "go-terms",
"interpro-identifiers", "taxonomy", "taxonomy-itsonedb",
"taxonomy-itsunite", "taxonomy-lsu", and "taxonomy-ssu". Note that
depending on the particular analysis type, pipeline version etc., not all
functional results will be available. (By default: \code{get.func = FALSE})}

\item{output}{A single character value specifying the format of an output.
Must be one of the following options: "TreeSE", "list", or "phyloseq".
(By default: \code{output = "TreeSE"})}

\item{use.cache}{A single boolean value specifying whether to use the
MGnify local caching system to speed up searching. It is highly
recommended that this be enabled. Note that files are downloaded to local system
when they are fetched from the database. The files are not removed meaning
that the local storage can include additional files after the run even though
\code{use.cache = FALSE} was specified. (By default: \code{use.cache = TRUE})}

\item{verbose}{A single boolean value to specify whether to show
the progress bar. (By default: \code{verbose = TRUE})}

\item{...}{optional arguments:
\itemize{
  \item{taxa.su}{ A single character value specifying which taxa subunit
  results should be selected? Currently, taxonomy assignments in the
  MGnify pipelines rely on rRNA matches to existing databases
  (GreenGenes and SILVA), with later pipelines checking both the SSU and
  LSU portions of the rRNA sequence. \code{taxa.su} allows then selection
  of either the Small subunit (SSU) or Large subunit results
  in the final \code{TreeSummarizedExperiment} object. Older pipeline
  versions do not report results for both subunits, and thus for some
  accessions this value will have no effect.}

  \item{get.tree}{ A single boolean value specifying whether to include
  available phylogenetic trees in the \code{TreeSummarizedExperiment}
  object. (By default: \code{get.tree = TRUE})}

  \item{as.df}{ A single boolean value enabled when \code{output = "list"}.
  The argument specifies whether return functional data as a named list
  (one entry per element in the output list) of data.frames,
  with each data.frame containing results for all requested accessions.
  If \code{FALSE}, The function returns a list of lists, each element
  consisting of results for a single accession.
  (By default: \code{as.df = TRUE})}

  \item{bulk.dl}{ A single boolean value specifying should
  MGnifyR attempt to speed things up by downloading
  relevant studies TSV results and only extracting the required columns,
  rather than using the JSONAPI interface. When getting results where
  multiple accessions share the same study, this option may result in
  significantly faster processing. However, there appear to be (quite a few)
  cases in the database where the TSV result columns do NOT match the
  expected accession names. This will hopefully be fixed in the future,
  but for now \code{bulk.dl} defaults to TRUE. When it does work, it can
  be orders of magnitude more efficient.
  (By default: \code{buld_dl = TRUE})}

}}
}
\value{
If only metagenomic data is retrieved, the result is returned in
\code{TreeSummarizedExperiment} object by default. The result can also be
returned as a \code{phyloseq} object or as a list of \code{data.frames}.
Note that \code{phyloseq} object can include only one phylogenetic tree
meaning that some taxa might be lost when data is subsetted based on tree.

When functional data is retrieved in addition to metagenomic data, the result
is returned as a \code{MultiAssayExperiment} object. Other options are a list
containing \code{phyloseq} object and \code{data.frames} or just
\code{data.frames}.

Functional data can be returned as a \code{MultiAssayExperiment} object or
as a list of \code{data.frames}.
}
\description{
Get functional and/or taxonomic information for a list of accessions
}
\details{
Given a set of analysis accessions and collection of annotation types,
the function queries the MGNify API and returns the results.
}
\examples{
\dontrun{
# Get OTU tables as TreeSE
tse <- getResult(mg, accession_list, get.func=FALSE, get.taxa=TRUE)

# Get functional data along with OTU tables as MAE
mae <- getResult(mg, accession_list, get.func=TRUE, get.taxa=TRUE)

# Get same data as list
analyses_res <- getResult(
    mg, accession_list, get.func=TRUE, get.taxa=TRUE, output = "list",
    as.df=T, use.cache = T, bulk.dl = F)
}

}
