% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_client}
\alias{mgnify_client}
\title{MGnify API client object.}
\usage{
mgnify_client(username = NULL, password = NULL, usecache = F,
  cachedir = NULL)
}
\description{
Acts as a simple container encapsulating API info (\code{baseurl}), user parameters (\code{authtok}) and cache
locations (\code{cache_dir})
}
\section{Slots}{

\describe{
\item{\code{baseurl}}{Web address of the MGnify JSON api - including version. Required}

\item{\code{authtok}}{The MGnify API supports authentication for users by way of \code{authtoken}. This is that token, althouth it's not used in anger at the moment.}

\item{\code{cache_dir}}{To reduce load on the server, and speed up repeated data processing, JSON calls may be cached locally, and
stored in \code{cache_dir}.}
}}

