% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_get_x_for_y}
\alias{mgnify_get_x_for_y}
\title{\code{JSONAPI} path for child elements}
\usage{
mgnify_get_x_for_y(client, x, typeX, typeY, usecache = F)
}
\arguments{
\item{client}{MGnifyR client API object}

\item{x}{Accession ID \code{char} of parent object}

\item{typeX}{Type of accession \code{x}}

\item{typeY}{Type of child object to return}
}
\value{
\code{char} complete url to access the result. Note this query is not run from here - just the URL is returned
}
\description{
\code{mgnify_get_x_for_y} determines the location of \code{typeY} child objects of \code{x} (type \code{typeX})
}
\details{
This helper function, principally intended to be used internally, is used to match up related objects within the path. The inherently
unhierarchical nature of the MGnify API makes it a bit inconsistent. This function acts as a quick way to determine how to get from
one type to another, without having to special case within the code.
}
\examples{
cl <- new("mgnify_client")
mgnify_get_x_for_y(cl, "MGYS00005126", "studies", "samples")
}
