% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\name{mgnify_query}
\alias{mgnify_query}
\title{#' Search MGnify database for studies, samples and runs}
\usage{
mgnify_query(client, qtype = "samples", accession = NULL,
  asDataFrame = F, ...)
}
\arguments{
\item{accession}{Either a single known MGnify accession identifier (of type \code{qtype}), or a list/vector
of accessions to query.}

\item{\code{mgnify_client}}{instance}

\item{\code{qtype}}{Type of objects to query. One of \code{studies},\code{samples},\code{runs} or
\code{analyses}}
}
\description{
\code{mgnify_query} is a flexible query function, harnessing the full power of the JSONAPI MGnify
search filters. Can be used for both metadata retrieval and
}
