% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allAccessors.R
\name{MgnifyClient-accessors}
\alias{MgnifyClient-accessors}
\alias{url,MgnifyClient-method}
\alias{authTok,MgnifyClient-method}
\alias{cacheDir,MgnifyClient-method}
\alias{warnings,MgnifyClient-method}
\alias{useMemCache,MgnifyClient-method}
\alias{memCache,MgnifyClient-method}
\alias{clearCache,MgnifyClient-method}
\alias{url<-,MgnifyClient-method}
\alias{authTok<-,MgnifyClient-method}
\alias{cacheDir<-,MgnifyClient-method}
\alias{warnings<-,MgnifyClient-method}
\alias{useMemCache<-,MgnifyClient-method}
\alias{memCache<-,MgnifyClient-method}
\alias{clearCache<-,MgnifyClient-method}
\title{Access and mutate slots of MgnifyClient object.}
\usage{
\S4method{url}{MgnifyClient}(x)

\S4method{authTok}{MgnifyClient}(x)

\S4method{cacheDir}{MgnifyClient}(x)

\S4method{warnings}{MgnifyClient}(x)

\S4method{useMemCache}{MgnifyClient}(x)

\S4method{memCache}{MgnifyClient}(x)

\S4method{clearCache}{MgnifyClient}(x)

\S4method{url}{MgnifyClient}(x) <- value

\S4method{authTok}{MgnifyClient}(x) <- value

\S4method{cacheDir}{MgnifyClient}(x) <- value

\S4method{warnings}{MgnifyClient}(x) <- value

\S4method{useMemCache}{MgnifyClient}(x) <- value

\S4method{memCache}{MgnifyClient}(x) <- value

\S4method{clearCache}{MgnifyClient}(x) <- value
}
\arguments{
\item{x}{MgnifClient object}

\item{value}{A value to be added to a specified slot.}
}
\value{
A value of a certain slot.
}
\description{
Access and mutate slots of MgnifyClient object.
}
\details{
Accessors and mutators of MgnifyClient object.
}
\examples{
mg <- MgnifyClient()
url(mg)

}
