% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/allAccessors.R
\name{databaseUrl}
\alias{databaseUrl}
\alias{authTok}
\alias{useCache}
\alias{cacheDir}
\alias{showWarnings}
\alias{useMemCache}
\alias{memCache}
\alias{clearCache}
\alias{verbose}
\alias{databaseUrl<-}
\alias{authTok<-}
\alias{useCache<-}
\alias{cacheDir<-}
\alias{showWarnings<-}
\alias{useMemCache<-}
\alias{memCache<-}
\alias{clearCache<-}
\alias{verbose<-}
\alias{MgnifyClient-accessors}
\alias{databaseUrl,MgnifyClient-method}
\alias{authTok,MgnifyClient-method}
\alias{useCache,MgnifyClient-method}
\alias{cacheDir,MgnifyClient-method}
\alias{showWarnings,MgnifyClient-method}
\alias{useMemCache,MgnifyClient-method}
\alias{memCache,MgnifyClient-method}
\alias{clearCache,MgnifyClient-method}
\alias{verbose,MgnifyClient-method}
\alias{databaseUrl<-,MgnifyClient-method}
\alias{authTok<-,MgnifyClient-method}
\alias{useCache<-,MgnifyClient-method}
\alias{cacheDir<-,MgnifyClient-method}
\alias{showWarnings<-,MgnifyClient-method}
\alias{useMemCache<-,MgnifyClient-method}
\alias{memCache<-,MgnifyClient-method}
\alias{clearCache<-,MgnifyClient-method}
\alias{verbose<-,MgnifyClient-method}
\title{MgnifyClient accessors and mutators}
\usage{
databaseUrl(x)

authTok(x)

useCache(x)

cacheDir(x)

showWarnings(x)

useMemCache(x)

memCache(x)

clearCache(x)

verbose(x)

databaseUrl(x) <- value

authTok(x) <- value

useCache(x) <- value

cacheDir(x) <- value

showWarnings(x) <- value

useMemCache(x) <- value

memCache(x) <- value

clearCache(x) <- value

verbose(x) <- value

\S4method{databaseUrl}{MgnifyClient}(x)

\S4method{authTok}{MgnifyClient}(x)

\S4method{useCache}{MgnifyClient}(x)

\S4method{cacheDir}{MgnifyClient}(x)

\S4method{showWarnings}{MgnifyClient}(x)

\S4method{useMemCache}{MgnifyClient}(x)

\S4method{memCache}{MgnifyClient}(x)

\S4method{clearCache}{MgnifyClient}(x)

\S4method{verbose}{MgnifyClient}(x)

\S4method{databaseUrl}{MgnifyClient}(x) <- value

\S4method{authTok}{MgnifyClient}(x) <- value

\S4method{useCache}{MgnifyClient}(x) <- value

\S4method{cacheDir}{MgnifyClient}(x) <- value

\S4method{showWarnings}{MgnifyClient}(x) <- value

\S4method{useMemCache}{MgnifyClient}(x) <- value

\S4method{memCache}{MgnifyClient}(x) <- value

\S4method{clearCache}{MgnifyClient}(x) <- value

\S4method{verbose}{MgnifyClient}(x) <- value
}
\arguments{
\item{x}{A \code{MgnifyClient} object.}

\item{value}{A value to be added to a certain slot.}
}
\value{
A value of MgnifyClient object or nothing.
}
\description{
MgnifyClient accessors and mutators
}
\details{
These functions are for fetching and mutating slots of
\code{MgnifyClient} object.
}
\examples{
mg <- MgnifyClient()

databaseUrl(mg)
showWarnings(mg) <- FALSE

}
