% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR-package.R
\docType{package}
\name{MGnifyR-package}
\alias{MGnifyR}
\alias{MGnifyR-package}
\title{MGnifyR: What the Package Does (Title Case)}
\description{
Utility package to facitliate easy interogation and analysis of EBI Mgnify pipeline results in R.
}
\details{
See the vignette for a general introduction to this package, [About MGnify](https://www.ebi.ac.uk/metagenomics/about) for general MGnify information, and
[API documentation](https://emg-docs.readthedocs.io/en/latest/api.html) for details about the JSONAPI implementation,
}
\keyword{internal}
