% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\docType{package}
\name{MGnifyR-package}
\alias{MGnifyR-package}
\title{\code{MGnifyR} Package.}
\description{
\code{MGnifyR} implements an interface to MGnify database. See the vignette
for a general introduction to this package,
[About MGnify](https://www.ebi.ac.uk/metagenomics/about) for general MGnify
information, and
[API documentation](https://emg-docs.readthedocs.io/en/latest/api.html) for
details about the JSONAPI implementation,
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
