% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGnifyR.R
\docType{_PACKAGE}
\name{MGnifyR-package}
\alias{MGnifyR-package}
\alias{MGnifyR}
\title{\code{MGnifyR} Package.}
\description{
\code{MGnifyR} implements an interface to the EBI MGnify database.
See the vignette for a general introduction to this package.
\href{https://www.ebi.ac.uk/metagenomics/about}{About MGnify} for general MGnify
information, and
\href{https://emg-docs.readthedocs.io/en/latest/api.html}{API documentation}
for details about the JSONAPI implementation,
}
\seealso{
\link[TreeSummarizedExperiment:TreeSummarizedExperiment-class]{TreeSummarizedExperiment} class
}
